fix(kubernetes): Run job stage does not always add suffix (#4509)
* test(kubernetes): Fix class names in test

I just noticed that a couple of test classes don't match their file
name; fix these.

* fix(kubernetes): Run job stage does not always add suffix

The Kubernetes run job stage currently always adds a random suffix
to the job name that it's deploying. This is so that we can re-run
the same job and not run into issues trying to replace a job.

There is already functionality built into Kubernetes for handling this;
by setting the generateName, Kubernetes will generate a random suffix
and append it.

This commit removes the logic to automatically add a suffix from all
jobs from Spinnaker; users should instead use a generateName if they
want this functionality. This now allows users who do not want a
random suffix to just use the name field and deploy the same job name
every time. This *does* mean that upon upgrading to this version, users
who want the same behavior as before will need to change their manifests
to use generateName instead of name.

We also force the use of the strategy.spinnaker.io/recreate deploy
strategy. This is because neither 'kubectl apply' nor 'kubectl replace'
will have the desired effect for a run job stage. In both of these cases,
if the job has not changed, the operation will be a successful no-op but
the job won't re-run, which will be confusing. If the job has changed, the
operation will fail because jobs are immutable. By forcing use of the
recreate strategy we'll ensure that a new job is created each time the
stage is run, and that there will be a new execution of the job each time
the stage is run.

* fix(test): Remove unnecessary call to set account to V2

V2 is the default now!