refactor(kubernetes): A few assorted code cleanup commits (#4512)
* refactor(kubernetes): A few minor improvements to KubernetesKind

Use the new NonnullByDefault annotation, and remove a bunch of
now redundant Nonnull annotations.

Also, simplify fromString to use a Guava splitter, which avoids
needing to use an array, and can simplify the logic. This is the
only part of this PR that isn't a simple rewrite of one function
to another, but the modified function here has a number of good
tests on it.

* refactor(kubernetes): Standardize on Nonnull

There was only a single place using a different annotation; change
it.

* refactor(kubernetes): Replace StringUtils.isEmpty with Guava

New code has been using isNullOrEmpty for a while; to reduce the
overhead for people reading the code let's standardize on the guava
version. Guava also has the advantage of encouraging nullToEmpty
to eliminate nulls as soon as they arrive and allow later checks
to avoid null-safety issues.

* refactor(kubernetes): Replace StringUtils.isNotEmpty with Guava

The motivation for this is in the last commit, but to make it
easier to read, did the isNotEmpty change in a separate commit.

In one case it was really easy to instead replace with nullToEmpty,
so I did that instead.

* refactor(kubernetes): Replace Collections.empty... with Guava

In a lot of places we use Collections.emptyList(), emptyMap(),
While it's not in the name of the function, the collection that
is returned is actually immutable.

Now that we've started adding lots of uses of Guava's immutable
collections in the Kubernetes provider, it makes sense for these
to also use the Guava types. This has a few advantages:
* It reduces the burden on new contributors to figure out which
of the two ways of doing the same thing they should use.
* The more we have Immutable collections floating around, the
more likely people will find that code in examples and use them
in their contributions.
* It's much more clear that the returned type is immutable
* In some cases, this allows us to also upgrade the return type
of the entire function to an immutable collection (when the only
return path was the one we just updated)

This change is safe because, as noted above, the collections were
already immutable. (Wheras changing a mutable collection to immutable
requires some care that it's not being mutated somewhere downstream.)

* refactor(kubernetes): Replace singleton collections with Guava

The rationale is in the last commit; this commit does the same
thing the last commit did with empty collections and does it for
singleton collections. Again, as these are already immutable, the
change is safe and actually makes it easier in some cases for the
caller to realize the returned collection is immutable.

* refactor(kubernetes): Remove unused functions

I removed the last place these two functions were used last week;
I noticed they were still here and unused while making the last
commits; just remove them entirely.