refactor(core): Clean up some dependencies (#4519)
I'm trying to reduce the amount of things on the compile path by
cleaning up dependencies. This cuts a few things from the
clouddriver-kubernetes-v2 module, and also adds a few things to
other modules that are used but not explicity included. (This is
to prepare for an upcoming kork bump where these things will not
be automatically included for everyone.)

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>