feat(dns): remove the DNS caching changes (#605)
Firstly, this won't compile under Java 11. I believe this class is available at runtime, so you could get it to work via reflection instead, but...

This doesn't feel like something an application developer should be setting. If you're concerned about DNS caching times, either set the TTL to 0 at the DNS level or configure this via JVM properties. Disabling DNS caching for the entire JVM is something that an operator should make a decision about, not us.

Obviously I'm happy to wait to submit this until you've had time to set the appropriate JVM properties in your deployments, but I figured I'd open a PR to start the discussion...