chore(jinjava): add test for jinjava breaks (#3608)
Various Jinjava issues have been problemetic for us in the past in how they affect MPTv1.
As we are in the process of upgrading to latest Jinjava we discovered this issue: https://github.com/HubSpot/jinjava/issues/429

This tests is reasonably representative of various issues we've seen including the issue above. Adding it so we have some sense of validation
when we bump Jinjava version

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>