feat(core): Merge collections when merging stage outputs (#3651)
* feat(core): Merge lists when merging stage outputs

Currently, when merging stage outputs, if equal keys are found, one of them will overwrite the other one. This can be a problem if you, for instance, run a baking pipeline that produces several different bakes, and you start this baking pipeline as a pipeline stage from a parent pipeline (i.e. not as a trigger). Each bake stage will produce a `deploymentDetails: [ bakeContext ... ]` and put this into the stage outputs. When the pipeline stage completes, only the last of these `deploymentDetails` will be present in the execution context of the parent pipeline.

This PR fixes this by checking if the duplicated key contains lists, and if so, it will concatenate them instead of overwriting. For everything else the behaviour will be like before.

* Same behaviour for all `Collection`s, not just `List`s

Add test data for `Set` and `distinct()`

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>