fix(fiat): only attempt to sync permissions when fiat is enabled (#830)
* fix(fiat): only attempt to sync permissions when fiat is enabled

In [this PR](https://github.com/spinnaker/front50/pull/619), we added logic to sync roles after an application is created. Let's only attempt to sync roles if the Fiat client is actually enabled. Elsewhere in Front50, such as in ApplicationPermissionsService.syncUsers, we only attempt to call Fiat to sync these roles if FiatClientConfigurationProperties is enabled, if FiatService is present, and if FiatConfigurationProperties.roleSync is enabled. Let's mirror this logic in the ApplicationsController.

* feat(config): add FiatStatus bean

* refactor(fiat): consume FiatStatus in ApplicationsController

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>