config(core): Push Halyard default configs down to core (#625)
As with the other microservices, let's push defaults that were
being set by Halyard down to the core so that the defaults are
the same for everyone.

Halyard was setting the redis connection based on the services.redis
value; push that down to the base. Users who were manually setting
this before will not be affected as they will override this value;
users who did not set this at all before will still fall back to
the localhost:6379 which is the fallback at the point we read it.

Halyard was also setting the configDir to /opt/packer/config. The
default in the base config was placeholder /some/path/to/config/packer;
replace this with what Halyard was setting. I suppose that if someone
was actually putting their templates in /some/path/to/config/packer then
this would break them until they explicitly set that value, but I can't
imagine that's actually the case.